+++ src/client/game/entity/entity.js
@@ -1,0 +1,12 @@
+import { Glyph } from 'glyph';
+
+export function Entity(properties = {}) {
+
+    return Object.assign(Glyph(properties), {
+        getPosition() {
+
+        }
+
+
+    });
+}
