+++ src/client/game/entity/glyph.js
@@ -1,0 +1,13 @@
+import { GC_BLACK, GC_WHITE } from '../core/constants';
+
+const background    = GC_BLACK;
+const character     = ' ';
+const foreground    = GC_WHITE;
+
+export function Glyph(properties = { background, character, foreground }) {
+    return Object.assign({}, {
+        getBackground: () => properties.background || background,
+        getCharacter: () => properties.character || character,
+        getForeground: () => properties.foreground || foreground
+    });
+}
