+++ src/client/game/state/lose.js
