+++ test/game/state/title_spec.js
@@ -3,35 +3,35 @@ import sinonChai    from 'sinon-chai';
 import sinon        from 'sinon';
 
 import { VK_RETURN }            from 'rot-js';
-import { LoseState }            from '../../../src/client/game/states/lose';
-import { GE_KEYDOWN, GS_TITLE } from '../../../src/client/game/core/constants';
+import { TitleState }           from '../../../src/client/game/state/title';
+import { GE_KEYDOWN, GS_PLAY }  from '../../../src/client/game/core/constants';
 
 chai.should();
 chai.use(sinonChai);
 
 const expect = chai.expect;
 
-describe('LoseState', () => {
+describe('TitleState', () => {
     describe('State transitions', () => {
-        it('Transition to title state on return keydown', () => {
+        it('Transition to play state on return keydown', () => {
             const game = {
                 switchTo(key) {}
             };
 
             sinon.spy(game, 'switchTo');
 
-            const state = LoseState(game);
+            const state = TitleState(game);
 
             state.handle(GE_KEYDOWN, VK_RETURN);
 
             game.switchTo.should.have.been.calledOnce;
-            game.switchTo.should.have.been.calledWith(GS_TITLE);
+            game.switchTo.should.have.been.calledWith(GS_PLAY);
         });
     });
 
     describe('Screen rendering', () => {
-        it('Render return to title prompt', () => {
-            const state = LoseState({
+        it('Render game title', () => {
+            const state = TitleState({
                 switchTo(key) {}
             });
 
@@ -43,7 +43,10 @@ describe('LoseState', () => {
 
             state.render(display);
 
-            display.drawText.should.have.been.calledWith(24, 21, 'Press [Enter] to return to title');
+            display.drawText.should.have.been.calledThrice;
+            display.drawText.should.have.been.calledWith(32, 9, '%c{white}The Dungeons of');
+            display.drawText.should.have.been.calledWith(37, 11, '%c{yellow}GLOOM');
+            display.drawText.should.have.been.calledWith(28, 21, 'Press [Enter] to start');
         });
     });
 });
